Add api call

+++ app/src/main/java/com/example/metaweather/data/entity/ResponseLocation.kt
@@ -16,21 +16,21 @@ data class ResponseLocation(
 )
 
 data class ConsolidatedWeather(
-    val id: Int,
+    val id: Double,
     val weather_state_name: String,
     val weather_state_abbr: String,
     val wind_direction_compass: String,
     val created: String,
     val applicable_date: String,
-    val min_temp: Int,
-    val max_temp: Int,
-    val the_temp: Int,
-    val wind_speed: Float,
-    val wind_direction: Float,
-    val air_pressure: Float,
-    val humidity: Float,
-    val visibility: Float,
-    val predictability: Int
+    val min_temp: Double,
+    val max_temp: Double,
+    val the_temp: Double,
+    val wind_speed: Double,
+    val wind_direction: Double,
+    val air_pressure: Double,
+    val humidity: Double,
+    val visibility: Double,
+    val predictability: Double
 )
 
 data class Parent(
